Restructure balance-overrides: extract balance/ submodule, drop config strategies
Moves balance override logic into dedicated `balance/` and `balance/aave` submodules,
gives the `Detector` its own cache and web3 handle, and renames the main struct to
`StateOverrides`. Deletes `configs::balance_overrides` (hardcoded per-token strategies
and the `autodetect` toggle) — detection is now always automatic.
