+++ crates/price-estimation/src/config/price_estimation.rs
@@ -10,19 +10,11 @@ pub trait BalanceOverridesConfigExt {
 
 impl BalanceOverridesConfigExt for BalanceOverridesConfig {
     fn init(&self, web3: ethrpc::Web3) -> Arc<dyn StateOverriding> {
-        Arc::new(balance_overrides::BalanceOverrides {
-            hardcoded: self.token_overrides.inner().clone(),
-            detector: self.autodetect.then(|| {
-                (
-                    balance_overrides::detector::Detector::new(
-                        web3.clone(),
-                        self.probing_depth,
-                        self.detection_timeout,
-                    ),
-                    std::sync::Mutex::new(cached::SizedCache::with_size(self.cache_size)),
-                )
-            }),
-            web3: Some(web3),
-        })
+        Arc::new(balance_overrides::StateOverrides::with_config(
+            web3,
+            self.probing_depth,
+            self.detection_timeout,
+            self.cache_size,
+        ))
     }
 }
